pengenalan state management menggunakan Redux atau Context
