+++ src/container/Pages/Product/CardProduct/Counter.jsx
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 
-class CardProduct extends Component {
+class Counter extends Component {
     state = {
         order: 3,
         // name: "Rian Mupti Jauhari"
     }
 
     onHandleChange = (newValue) => {
-        this.props.onChange(newValue)
+        this.props.onCounterChange(newValue)
     }
 
     handlePlus = () => {
@@ -27,20 +27,16 @@ class CardProduct extends Component {
             })
         }
     }
+
     render() {
         return (
-            <div className="card">
-                <img src="" alt="image-product" />
-                <p className="product-title">Daging ayam rasa sapi</p>
-                <p className="product-price">Rp 340,000</p>
-                <div className="counter">
-                    <button className="minus" onClick={this.handleMinus}>-</button>
-                    <input type="text" value={this.state.order} />
-                    <button className="plus" onClick={this.handlePlus}>+</button>
-                </div>
+            <div className="counter">
+                <button className="minus" onClick={this.handleMinus}>-</button>
+                <input type="text" value={this.state.order} />
+                <button className="plus" onClick={this.handlePlus}>+</button>
             </div>
         )
     }
 }
 
-export default CardProduct;
+export default Counter;
